requirements.txt
torch>=1.10
torchvision>=0.11
numpy
Pillow
pytorch-msssim
kornia
lpips
